move converter into processing.pre
