--- app/openqasm3/parser.py
@@ -1,16 +1,0 @@
-"""
-Extended parsing of abstract syntax trees with support for `OPENQASM 2.x`.
-"""
-
-from openqasm3.ast import Program
-
-from app.converter import parse_to_openqasm3
-
-
-def leqo_parse(qasm: str) -> Program:
-    """Parse an openqasm2 or openqasm3 string into an ast (:class:`~openqasm3.ast.Program`).
-
-    :param qasm: The qasm string to parse
-    :return: The parse ast
-    """
-    return parse_to_openqasm3(qasm)
